refactor: split rerun code so .rrd utilities work without rerun-sdk
Downstream projects that generate .rrd files via the C++ rerun library
can now use bencher's viewing/serving functions without installing
rerun-sdk (which requires NumPy 2).

New utils_rrd.py (no rerun-sdk dependency):
- rrd_to_pane, publish_and_view_rrd, rrd_file_to_pane

utils_rerun.py now only contains SDK-dependent functions:
- capture_rerun_rrd, rerun_to_pane, capture_rerun_window
